xhtml_parser

Module node

Node representation and manipulation in an XML document.

This module defines the Node struct, which represents a node in an XML document. It provides methods to access node properties, navigate the document tree, and retrieve attributes and text content.

The Node struct is designed to work with a Document, which contains the XML data and node information.

Example

use xhtml_parser::Document;
use xhtml_parser::Node;

let xml_data = b"<root><child>Text</child></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let child_node = root_node.first_child().unwrap();

assert_eq!(child_node.tag_name(), "child");

let child_node = child_node.first_child().unwrap();

assert_eq!(child_node.text().unwrap(), "Text");
assert!(!child_node.is_element());
assert!(child_node.is_text());

Features

Access to node attributes and namespaces

Navigation through the document tree (children, siblings)

Support for different node types (elements, text, etc.)

Iteration over node children

Note

This module is part of the xhtml_parser crate and is designed to work with XML documents.

Structs

Node

Represents a node in an XML document.

NodeChildren

Iterator over node children.