xhtml_parser::node

Struct Node

pub struct Node<'xml> {
    pub idx: NodeIdx,
    pub node_info: &'xml NodeInfo,
    pub doc: &'xml Document,
}

Represents a node in an XML document.

Node contains metadata about the node, such as its index, type, and position in the document. It provides methods to access the nodeโs tag name, text content, attributes, and navigation through the document tree.

Fields
idx: NodeIdxnode_info: &'xml NodeInfodoc: &'xml Document
Implementations

impl<'xml> Node<'xml>

pub fn new(idx: NodeIdx, node_info: &'xml NodeInfo, doc: &'xml Document) -> Self

Creates a new Node instance.

Arguments

idx: The index of the node in the document.

node_info: A reference to the NodeInfo containing metadata about the node.

doc: A reference to the Document containing the XML data.

pub fn idx(&self) -> NodeIdx

Returns the index of the node in the document.

pub fn tag_name(&self) -> &str

Returns the tag name of the node. If the node is not an element, it returns an empty string.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child>Text</child></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let tag_name = root_node.tag_name();
assert_eq!(tag_name, "root");

pub fn text(&self) -> &str

Returns the text content of the node. If the node is not a text node, it returns an empty string.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root>Text</root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let child_node = root_node.first_child().unwrap();
assert!(child_node.is_text());
let text_content = child_node.text();
assert_eq!(text_content, "Text");

pub fn attributes(&self) -> Attributes<'xml>

Returns a new Attributes iterator instance for this node.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
use xhtml_parser::attribute::Attributes;
let xml_data = b"<root name=\"The root\" id=\"1\">Text</root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let attributes: Vec<_> = root_node.attributes().collect();
assert_eq!(attributes.len(), 2);
assert_eq!(attributes[0].name(), "name");
assert_eq!(attributes[0].value(), "The root");
assert_eq!(attributes[1].name(), "id");
assert_eq!(attributes[1].value(), "1");

pub fn first_child(&self) -> Option<Node<'xml>>

Returns the first child of the node, if it exists, None otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let node = document.root().unwrap();    
assert_eq!(node.first_child().unwrap().tag_name(), "child1");

pub fn last_child(&self) -> Option<Node<'xml>>

Returns the last child of the node, if it exists, None otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let last_child = root_node.last_child().unwrap();
assert_eq!(last_child.tag_name(), "child2");

pub fn next_sibling(&self) -> Option<Node<'xml>>

Returns the next sibling of the node, if it exists, None otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let next_sibling = root_node.first_child().unwrap().next_sibling().unwrap();
assert_eq!(next_sibling.tag_name(), "child2");

pub fn prev_sibling(&self) -> Option<Node<'xml>>

Returns the previous sibling of the node, if it exists, None otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let prev_sibling = root_node.last_child().unwrap().prev_sibling().unwrap();
assert_eq!(prev_sibling.tag_name(), "child1");

pub fn children(&self) -> NodeChildren<'_>

Returns an iterator over the children of the node. If the node has no children, it returns an empty iterator.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let children: Vec<_> = root_node.children().collect();
assert_eq!(children.len(), 2);
assert_eq!(children[0].tag_name(), "child1");
assert_eq!(children[1].tag_name(), "child2");

pub fn is_root(&self) -> bool

Returns true if the node is the root node, false otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
assert!(root_node.is_root());

pub fn has_children(&self) -> bool

Returns true if the node has children, false otherwise.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
assert!(root_node.has_children());

pub fn is_element(&self) -> bool

Returns true if the node is a NodeType::Element, false otherwise.

pub fn is_text(&self) -> bool

Returns true if the node is a NodeType::Text, false otherwise.

pub fn get_node_type(&self) -> &NodeType

Returns the NodeType instance associated with this node.

pub fn find_child(&self, tag_name: &str) -> Option<Node<'xml>>

Finds a child node with the specified tag name. If the node has no children, it returns None.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
if let Some(child) = root_node.find_child("child2") {
    assert_eq!(child.tag_name(), "child2");
} else {
    panic!("Child node not found");
}

pub fn find_sibling(&self, tag_name: &str) -> Option<Node<'xml>>

Finds a sibling node with the specified tag name. If the node has no parent or no siblings, it returns None.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child1/><child2/></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let child_node = root_node.first_child().unwrap();
if let Some(sibling) = child_node.find_sibling("child2") {
    assert_eq!(sibling.tag_name(), "child2");
} else {
    panic!("Sibling node not found");
}

pub fn find_attribute(&self, name: &str) -> Option<String>

searches for an attribute by name and returns its value if found.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root name=\"value\">Text</root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
if let Some(value) = root_node.find_attribute("name") {
    assert_eq!(value, "value");
} else {
    panic!("Attribute not found");
}

pub fn parent(&self) -> Option<Node<'xml>>

Returns the parent node of this node, if it exists. If this node is the root node, it returns None.

Example

use xhtml_parser::document::Document;
use xhtml_parser::node::Node;
let xml_data = b"<root><child>Text</child></root>".to_vec();
let document = Document::new(xml_data).unwrap();
let root_node = document.root().unwrap();
let child_node = root_node.first_child().unwrap();
if let Some(parent) = child_node.parent() {
    assert_eq!(parent.tag_name(), "root");
} else {
    panic!("Child node has no parent");
}